+++ gazebo/gui/ModelAlign.cc
@@ -61,10 +61,10 @@ ModelAlign::~ModelAlign()
 void ModelAlign::Clear()
 {
   this->dataPtr->targetVis.reset();
-  this->dataPtr->userCamera.reset();
   this->dataPtr->scene.reset();
   this->dataPtr->node.reset();
   this->dataPtr->modelPub.reset();
+  this->dataPtr->userCmdPub.reset();
   this->dataPtr->selectedVisuals.clear();
   this->dataPtr->connections.clear();
   this->dataPtr->originalVisualPose.clear();
@@ -79,18 +79,26 @@ void ModelAlign::Init()
 
   rendering::UserCameraPtr cam = gui::get_active_camera();
   if (!cam)
-    return;
+  {
+    this->dataPtr->scene = rendering::get_scene();
+  }
+  else
+  {
+    this->dataPtr->scene = cam->GetScene();
+  }
 
-  if (!cam->GetScene())
-    return;
-
-  this->dataPtr->userCamera = cam;
-  this->dataPtr->scene = cam->GetScene();
+  if (!this->dataPtr->scene)
+  {
+    gzerr << "Unable to initialize Model Align tool, scene is NULL"
+        << std::endl;
+  }
 
   this->dataPtr->node = transport::NodePtr(new transport::Node());
   this->dataPtr->node->Init();
   this->dataPtr->modelPub =
       this->dataPtr->node->Advertise<msgs::Model>("~/model/modify");
+  this->dataPtr->userCmdPub =
+      this->dataPtr->node->Advertise<msgs::UserCmd>("~/user_cmd");
 
   this->dataPtr->initialized = true;
 }
@@ -289,6 +297,15 @@ void ModelAlign::AlignVisuals(std::vector<rendering::VisualPtr> _visuals,
 
     if (_publish)
       this->PublishVisualPose(vis);
+  }
+  // Register user command on server
+  if (_publish)
+  {
+    msgs::UserCmd userCmdMsg;
+    userCmdMsg.set_description(
+        "Align to [" + this->dataPtr->targetVis->GetName() + "]");
+    userCmdMsg.set_type(msgs::UserCmd::MOVING);
+    this->dataPtr->userCmdPub->Publish(userCmdMsg);
   }
 }