+++ gazebo/gui/model/ModelEditorEvents.cc
@@ -27,14 +27,14 @@ event::EventT<void ()> model::Events::newModelEditor;
 event::EventT<void ()> model::Events::exitModelEditor;
 event::EventT<void ()> model::Events::modelChanged;
 event::EventT<void (std::string)> model::Events::modelNameChanged;
-event::EventT<void (bool, bool, const math::Pose &, const std::string &)>
-    model::Events::modelPropertiesChanged;
+event::EventT<void (bool, bool)> model::Events::modelPropertiesChanged;
 
 event::EventT<void (std::string)> model::Events::saveModel;
 event::EventT<void ()> model::Events::newModel;
 
 event::EventT<void (std::string)> model::Events::nestedModelInserted;
 event::EventT<void (std::string)> model::Events::linkInserted;
+event::EventT<void (std::string)> model::Events::nestedModelRemoved;
 event::EventT<void (std::string)> model::Events::linkRemoved;
 event::EventT<void (std::string)> model::Events::requestLinkRemoval;
 event::EventT<void (std::string, std::string, std::string, std::string,