+++ gazebo/sensors/WideAngleCameraSensor.cc
@@ -110,9 +110,9 @@ void WideAngleCameraSensor::Init()
     this->camera->Init();
     this->camera->CreateRenderTexture(this->GetName() + "_RttTex");
 
-    math::Pose cameraPose = this->pose;
+    ignition::math::Pose3d cameraPose = this->pose;
     if (cameraSdf->HasElement("pose"))
-      cameraPose = cameraSdf->Get<math::Pose>("pose") + cameraPose;
+      cameraPose = cameraSdf->Get<ignition::math::Pose3d>("pose") + cameraPose;
 
     this->camera->SetWorldPose(cameraPose);
     this->camera->AttachToVisual(this->parentId, true);
@@ -179,7 +179,10 @@ bool WideAngleCameraSensor::UpdateImpl(bool _force)
 
     for (; !this->dataPtr->hfovCmdQueue.empty();
         this->dataPtr->hfovCmdQueue.pop())
-      this->camera->SetHFOV(math::Angle(this->dataPtr->hfovCmdQueue.front()));
+    {
+      this->camera->SetHFOV(ignition::math::Angle(
+            this->dataPtr->hfovCmdQueue.front()));
+    }
 
     msgs::CameraLens msg;